Move qutip dims conversion methods to qarray.py
