+++ dynamiqs/qarrays/dense_qarray.py
@@ -11,7 +11,14 @@ from qutip import Qobj
 
 from .._utils import _is_batched_scalar
 from .layout import Layout, dense
-from .qarray import QArray, QArrayLike, _asjaxarray, _in_last_two_dims, isqarraylike
+from .qarray import (
+    QArray,
+    QArrayLike,
+    _asjaxarray,
+    _dims_to_qutip,
+    _in_last_two_dims,
+    isqarraylike,
+)
 
 __all__ = ['DenseQArray']
 
@@ -25,22 +32,6 @@ def _dense_to_qobj(x: DenseQArray) -> Qobj | list[Qobj]:
     else:
         dims = _dims_to_qutip(x.dims, x.shape)
         return Qobj(x, dims=dims)
-
-
-def _dims_to_qutip(dims: tuple[int, ...], shape: tuple[int, ...]) -> list:
-    dims = list(dims)
-    if shape[-1] == 1:  # [[3], [1]] or [[3, 4], [1, 1]]
-        dims = [dims, [1] * len(dims)]
-    elif shape[-2] == 1:  # [[1], [3]] or [[1, 1], [3, 4]]
-        dims = [[1] * len(dims), dims]
-    elif shape[-1] == shape[-2]:  # [[3], [3]] or [[3, 4], [3, 4]]
-        dims = [dims, dims]
-    return dims
-
-
-def _dims_from_qutip(dims: list) -> tuple[int, ...]:
-    dims = np.max(dims, axis=0)
-    return tuple([d.item() for d in dims])
 
 
 class DenseQArray(QArray):
